Move MPS device check to cheetah.utils
